+++ src/whl2conda/cli/config.py
@@ -1,4 +1,4 @@
-#  Copyright 2023 Christopher Barber
+#  Copyright 2023-2024 Christopher Barber
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -19,6 +19,7 @@ whl2conda config subcommand implementation
 from __future__ import annotations
 
 import argparse
+import json
 import sys
 from http.client import HTTPException
 from pathlib import Path
@@ -29,6 +30,7 @@ from ..api.stdrename import update_renames_file
 from .common import add_markdown_help, dedent
 from ..impl.pyproject import add_pyproject_defaults
 from ..api.stdrename import user_stdrenames_path
+from ..settings import settings, _fromidentifier
 
 __all__ = ["config_main"]
 
@@ -76,6 +78,69 @@ def config_main(
             %(const)s.
             """),
     )
+
+    settings_opts = parser.add_argument_group(
+        "User settings options",
+        description=dedent(
+            """
+            These options allow you to view or modify persistent user settings
+            that affect the behavior of whl2conda. Note that all of these options
+            treat key dash and underscore in key names as equivalent (e.g. you
+            may use either `conda-format` or `conda_format`).
+            """
+        ),
+    )
+
+    # TODO - add a --describe option
+
+    settings_opts.add_argument(
+        "--set",
+        metavar=("<key>", "<value>"),
+        nargs=2,
+        help=dedent("""
+            Sets configuration parameter specified by <key> to given <value>.
+            For dictionary attributes (e.g. pypi-indexes), the <key> should be
+            of the form `<key>.<entry>` to set a specific entry in the table, e.g.
+            
+               whl2conda config --set pypi-indexes.acme https://acme.com/pypi
+               
+            Note that it is not currently possible to assign to entire dictionary.
+        """),
+    )
+    settings_opts.add_argument(
+        "--remove",
+        metavar="<key>",
+        help=dedent(
+            """
+            Unset user setting with given key. To remove an entry from a dictionary
+            use a key for the form '<key>.<entry>', e.g.
+            
+              whl2conda config --remove pypi-indexes.acme
+            """
+        ),
+    )
+
+    show_opts = settings_opts.add_mutually_exclusive_group()
+    show_opts.add_argument(
+        "--show",
+        metavar="<key>",
+        nargs="*",
+        help=dedent("""
+            Show user settings from local settings file. If any <key> is given,
+            then only the specified settings will be displayed, otherwise
+            all settings will be shown.
+            """),
+    )
+
+    show_opts.add_argument(
+        "--show-sources",
+        action="store_true",
+        help=dedent("""
+            The same as --show with no arguments but also displays location
+            of settings file.
+            """),
+    )
+
     parser.add_argument(
         "-n",
         "--dry-run",
@@ -94,6 +159,17 @@ def config_main(
             add_pyproject_defaults(parsed.generate_pyproject)
         except Exception as ex:  # pylint: disable=broad-exception-caught
             parser.error(str(ex))
+
+    if parsed.remove:
+        remove_user_setting(parsed.remove, dry_run=parsed.dry_run)
+
+    if parsed.set:
+        set_user_setting(*parsed.set, dry_run=parsed.dry_run)
+
+    if parsed.show is not None:
+        show_user_settings(parsed.show)
+    elif parsed.show_sources:
+        show_user_settings(show_sources=True)
 
 
 def update_std_renames(renames_file: Path, *, dry_run: bool) -> None:
@@ -122,6 +198,36 @@ def update_std_renames(renames_file: Path, *, dry_run: bool) -> None:
         print(f"Cannot download update: {ex}", file=sys.stderr)
         sys.exit(8)
     sys.exit(0)
+
+
+def save_settings(*, dry_run) -> None:
+    if dry_run:
+        print(f"Dry run: not saving {settings.settings_file}")
+    else:
+        settings.save()
+
+
+def set_user_setting(key: str, value: str, *, dry_run: bool = False) -> None:
+    settings.set(key, value)
+    save_settings(dry_run=dry_run)
+
+
+def show_user_settings(
+    keys: Sequence[str] = (),
+    show_sources: bool = False,
+) -> None:
+    if show_sources:
+        print(f"==> {settings.settings_file} <==")
+    if keys:
+        for key in keys:
+            print(f"{_fromidentifier(key)}: {json.dumps(settings.get(key))}")
+    else:
+        print(json.dumps(settings.to_dict(), indent=2))
+
+
+def remove_user_setting(key: str, *, dry_run: bool = False) -> None:
+    settings.unset(key)
+    save_settings(dry_run=dry_run)
 
 
 if __name__ == "__main__":  # pragma: no cover